requirements.txt
pycryptodomex==3.21.0
pytest==8.2.2
